+++ tests/unittest/_torch/visual_gen/test_ltx2_pipeline.py
@@ -21,12 +21,13 @@ from test_common.llm_data import llm_models_root
 
 from tensorrt_llm._torch.modules.linear import Linear
 from tensorrt_llm._torch.visual_gen.config import (
-    LTX2_FORCE_ONE_STAGE_ENV,
     AttentionConfig,
+    CacheDiTConfig,
     DiffusionModelConfig,
     PipelineComponent,
     VisualGenArgs,
 )
+from tensorrt_llm._torch.visual_gen.models.ltx2.pipeline_ltx2 import LTX2_FORCE_ONE_STAGE_ENV
 from tensorrt_llm._torch.visual_gen.pipeline_loader import PipelineLoader
 
 os.environ.setdefault("TLLM_DISABLE_MPI", "1")
@@ -732,8 +733,9 @@ class TestTwoStageLoRAFileLoading:
 class TestTwoStagePipelineVariantResolution:
     """Test that LTX2Pipeline.resolve_variant selects the correct class."""
 
-    def test_resolve_variant_returns_two_stage_when_configured(self):
+    def test_resolve_variant_returns_two_stage_when_configured(self, monkeypatch):
         """When both upsampler and LoRA paths are set, resolve_variant returns TwoStages."""
+        monkeypatch.delenv(LTX2_FORCE_ONE_STAGE_ENV, raising=False)
         from unittest.mock import MagicMock
 
         from tensorrt_llm._torch.visual_gen.models.ltx2.pipeline_ltx2 import LTX2Pipeline
@@ -742,6 +744,7 @@ class TestTwoStagePipelineVariantResolution:
         )
 
         config = MagicMock()
+        config.pretrained_config._name_or_path = ""
         config.extra_attrs = {
             "spatial_upsampler_path": "/fake/upsampler.safetensors",
             "distilled_lora_path": "/fake/lora.safetensors",
@@ -750,15 +753,16 @@ class TestTwoStagePipelineVariantResolution:
         result = LTX2Pipeline.resolve_variant(config)
         assert result is LTX2TwoStagesPipeline
 
-    def test_resolve_variant_honors_force_one_stage_pipeline(self):
-        """force_one_stage_pipeline should prevent promotion even when two-stage paths exist."""
+    def test_resolve_variant_honors_force_one_stage_env(self, monkeypatch):
+        """The env knob should prevent promotion even when two-stage paths exist."""
         from unittest.mock import MagicMock
 
         from tensorrt_llm._torch.visual_gen.models.ltx2.pipeline_ltx2 import LTX2Pipeline
 
+        monkeypatch.setenv(LTX2_FORCE_ONE_STAGE_ENV, "1")
         config = MagicMock()
+        config.pretrained_config._name_or_path = ""
         config.extra_attrs = {
-            "force_one_stage_pipeline": True,
             "spatial_upsampler_path": "/fake/upsampler.safetensors",
             "distilled_lora_path": "/fake/lora.safetensors",
         }
@@ -773,6 +777,7 @@ class TestTwoStagePipelineVariantResolution:
         from tensorrt_llm._torch.visual_gen.models.ltx2.pipeline_ltx2 import LTX2Pipeline
 
         config = MagicMock()
+        config.pretrained_config._name_or_path = ""
         config.extra_attrs = {}
 
         result = LTX2Pipeline.resolve_variant(config)
@@ -785,6 +790,7 @@ class TestTwoStagePipelineVariantResolution:
         from tensorrt_llm._torch.visual_gen.models.ltx2.pipeline_ltx2 import LTX2Pipeline
 
         config = MagicMock()
+        config.pretrained_config._name_or_path = ""
         config.extra_attrs = {"spatial_upsampler_path": "/fake/upsampler.safetensors"}
 
         result = LTX2Pipeline.resolve_variant(config)
@@ -792,9 +798,14 @@ class TestTwoStagePipelineVariantResolution:
 
 
 class TestLTX2ForceOneStageEnv:
-    """Test force-one-stage env-var behavior before pipeline construction."""
+    """Test force-one-stage env-var behavior during LTX2 variant selection."""
 
     def test_two_stage_auxiliary_paths_are_discovered_by_default(self, tmp_path, monkeypatch):
+        from tensorrt_llm._torch.visual_gen.models.ltx2.pipeline_ltx2 import LTX2Pipeline
+        from tensorrt_llm._torch.visual_gen.models.ltx2.pipeline_ltx2_two_stages import (
+            LTX2TwoStagesPipeline,
+        )
+
         monkeypatch.delenv(LTX2_FORCE_ONE_STAGE_ENV, raising=False)
         checkpoint_path = _write_minimal_ltx2_diffusers_checkpoint(tmp_path)
         upsampler_path = checkpoint_path / "ltx-2-spatial-upscaler-x2-1.0.safetensors"
@@ -805,23 +816,32 @@ class TestLTX2ForceOneStageEnv:
         args = VisualGenArgs(checkpoint_path=str(checkpoint_path))
         config = DiffusionModelConfig.from_pretrained(str(checkpoint_path), args=args)
 
+        assert LTX2Pipeline.resolve_variant(config) is LTX2TwoStagesPipeline
         assert config.extra_attrs["spatial_upsampler_path"] == str(upsampler_path)
         assert config.extra_attrs["distilled_lora_path"] == str(lora_path)
 
-    def test_force_one_stage_env_disables_auxiliary_discovery(self, tmp_path, monkeypatch):
+    def test_force_one_stage_env_skips_auto_discovery(self, tmp_path, monkeypatch):
+        from tensorrt_llm._torch.visual_gen.models.ltx2.pipeline_ltx2 import LTX2Pipeline
+
         monkeypatch.setenv(LTX2_FORCE_ONE_STAGE_ENV, "1")
         checkpoint_path = _write_minimal_ltx2_diffusers_checkpoint(tmp_path)
-        (checkpoint_path / "ltx-2-spatial-upscaler-x2-1.0.safetensors").touch()
-        (checkpoint_path / "ltx-2-19b-distilled-lora-384.safetensors").touch()
+        upsampler_path = checkpoint_path / "ltx-2-spatial-upscaler-x2-1.0.safetensors"
+        lora_path = checkpoint_path / "ltx-2-19b-distilled-lora-384.safetensors"
+        upsampler_path.touch()
+        lora_path.touch()
 
         args = VisualGenArgs(checkpoint_path=str(checkpoint_path))
         config = DiffusionModelConfig.from_pretrained(str(checkpoint_path), args=args)
 
-        assert config.extra_attrs["force_one_stage_pipeline"] is True
+        assert LTX2Pipeline.resolve_variant(config) is LTX2Pipeline
         assert "spatial_upsampler_path" not in config.extra_attrs
         assert "distilled_lora_path" not in config.extra_attrs
 
-    def test_force_one_stage_env_ignores_explicit_auxiliary_paths(self, tmp_path, monkeypatch):
+    def test_force_one_stage_env_prevents_promotion_with_explicit_auxiliary_paths(
+        self, tmp_path, monkeypatch
+    ):
+        from tensorrt_llm._torch.visual_gen.models.ltx2.pipeline_ltx2 import LTX2Pipeline
+
         monkeypatch.setenv(LTX2_FORCE_ONE_STAGE_ENV, "1")
         checkpoint_path = _write_minimal_ltx2_diffusers_checkpoint(tmp_path)
 
@@ -832,9 +852,30 @@ class TestLTX2ForceOneStageEnv:
         )
         config = DiffusionModelConfig.from_pretrained(str(checkpoint_path), args=args)
 
-        assert config.extra_attrs["force_one_stage_pipeline"] is True
-        assert "spatial_upsampler_path" not in config.extra_attrs
-        assert "distilled_lora_path" not in config.extra_attrs
+        assert config.extra_attrs["spatial_upsampler_path"] == "/fake/upsampler.safetensors"
+        assert config.extra_attrs["distilled_lora_path"] == "/fake/lora.safetensors"
+        assert LTX2Pipeline.resolve_variant(config) is LTX2Pipeline
+
+    def test_cache_dit_config_prevents_promotion_with_explicit_auxiliary_paths(
+        self, tmp_path, monkeypatch
+    ):
+        from tensorrt_llm._torch.visual_gen.models.ltx2.pipeline_ltx2 import LTX2Pipeline
+
+        monkeypatch.delenv(LTX2_FORCE_ONE_STAGE_ENV, raising=False)
+        checkpoint_path = _write_minimal_ltx2_diffusers_checkpoint(tmp_path)
+
+        args = VisualGenArgs(
+            checkpoint_path=str(checkpoint_path),
+            cache=CacheDiTConfig(),
+            spatial_upsampler_path="/fake/upsampler.safetensors",
+            distilled_lora_path="/fake/lora.safetensors",
+        )
+        config = DiffusionModelConfig.from_pretrained(str(checkpoint_path), args=args)
+
+        assert config.cache_backend == "cache_dit"
+        assert config.extra_attrs["spatial_upsampler_path"] == "/fake/upsampler.safetensors"
+        assert config.extra_attrs["distilled_lora_path"] == "/fake/lora.safetensors"
+        assert LTX2Pipeline.resolve_variant(config) is LTX2Pipeline
 
 
 class TestTwoStageUpsamplerBuildingBlocks:
